add base addon manager.
