+++ pkg/addonmanager/cloudevents/manager.go
@@ -2,11 +2,9 @@ package cloudevents
 
 import (
 	"context"
-	"fmt"
 	"time"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/client-go/dynamic"
 	"k8s.io/client-go/dynamic/dynamicinformer"
 	kubeinformers "k8s.io/client-go/informers"
@@ -18,9 +16,7 @@ import (
 	addoninformers "open-cluster-management.io/api/client/addon/informers/externalversions"
 	clusterv1client "open-cluster-management.io/api/client/cluster/clientset/versioned"
 	clusterv1informers "open-cluster-management.io/api/client/cluster/informers/externalversions"
-	workclientset "open-cluster-management.io/api/client/work/clientset/versioned"
 	workinformers "open-cluster-management.io/api/client/work/informers/externalversions"
-	workv1informers "open-cluster-management.io/api/client/work/informers/externalversions/work/v1"
 
 	"open-cluster-management.io/addon-framework/pkg/addonmanager"
 	"open-cluster-management.io/addon-framework/pkg/addonmanager/controllers/addonconfig"
@@ -29,7 +25,6 @@ import (
 	"open-cluster-management.io/addon-framework/pkg/addonmanager/controllers/cmaconfig"
 	"open-cluster-management.io/addon-framework/pkg/addonmanager/controllers/cmamanagedby"
 	"open-cluster-management.io/addon-framework/pkg/addonmanager/controllers/registration"
-	"open-cluster-management.io/addon-framework/pkg/agent"
 	"open-cluster-management.io/addon-framework/pkg/basecontroller/factory"
 	"open-cluster-management.io/addon-framework/pkg/index"
 	"open-cluster-management.io/addon-framework/pkg/utils"
@@ -37,35 +32,21 @@ import (
 	"open-cluster-management.io/sdk-go/pkg/cloudevents/work/source/codec"
 )
 
+// cloudeventsAddonManager is the implementation of AddonManager with
+// the base implementation and cloudevents options
 type cloudeventsAddonManager struct {
-	addonAgents  map[string]agent.AgentAddon
-	addonConfigs map[schema.GroupVersionResource]bool
-	config       *rest.Config
-	options      *CloudEventsOptions
-	syncContexts []factory.SyncContext
-}
-
-func (a *cloudeventsAddonManager) AddAgent(addon agent.AgentAddon) error {
-	addonOption := addon.GetAgentAddonOptions()
-	if len(addonOption.AddonName) == 0 {
-		return fmt.Errorf("addon name should be set")
-	}
-	if _, ok := a.addonAgents[addonOption.AddonName]; ok {
-		return fmt.Errorf("an agent is added for the addon already")
-	}
-	a.addonAgents[addonOption.AddonName] = addon
-	return nil
-}
-
-func (a *cloudeventsAddonManager) Trigger(clusterName, addonName string) {
-	for _, syncContex := range a.syncContexts {
-		syncContex.Queue().Add(fmt.Sprintf("%s/%s", clusterName, addonName))
-	}
+	*addonmanager.BaseAddonManagerImpl
+	options *CloudEventsOptions
 }
 
 func (a *cloudeventsAddonManager) Start(ctx context.Context) error {
+	config := a.GetConfig()
+	addonAgents := a.GetAddonAgents()
+	addonConfigs := a.GetAddonConfigs()
+	syncContexts := a.GetSyncContexts()
+
 	var addonNames []string
-	for key := range a.addonAgents {
+	for key := range addonAgents {
 		addonNames = append(addonNames, key)
 	}
 
@@ -73,8 +54,8 @@ func (a *cloudeventsAddonManager) Start(ctx context.Context) error {
 	// ManifestWork client that implements the ManifestWorkInterface and ManifestWork informer based on different
 	// driver configuration.
 	// Refer to Event Based Manifestwork proposal in enhancements repo to get more details.
-	_, config, err := cloudeventswork.NewConfigLoader(a.options.WorkDriver, a.options.WorkDriverConfig).
-		WithKubeConfig(a.config).
+	_, clientConfig, err := cloudeventswork.NewConfigLoader(a.options.WorkDriver, a.options.WorkDriverConfig).
+		WithKubeConfig(a.GetConfig()).
 		LoadConfig()
 	if err != nil {
 		return err
@@ -97,7 +78,7 @@ func (a *cloudeventsAddonManager) Start(ctx context.Context) error {
 		},
 	)
 
-	clientHolder, err := cloudeventswork.NewClientHolderBuilder(config).
+	clientHolder, err := cloudeventswork.NewClientHolderBuilder(clientConfig).
 		WithClientID(a.options.CloudEventsClientID).
 		WithSourceID(a.options.SourceID).
 		WithInformerConfig(10*time.Minute, workInformOption).
@@ -107,22 +88,22 @@ func (a *cloudeventsAddonManager) Start(ctx context.Context) error {
 		return err
 	}
 
-	kubeClient, err := kubernetes.NewForConfig(a.config)
+	kubeClient, err := kubernetes.NewForConfig(config)
 	if err != nil {
 		return err
 	}
 
-	dynamicClient, err := dynamic.NewForConfig(a.config)
+	dynamicClient, err := dynamic.NewForConfig(config)
 	if err != nil {
 		return err
 	}
 
-	addonClient, err := addonv1alpha1client.NewForConfig(a.config)
+	addonClient, err := addonv1alpha1client.NewForConfig(config)
 	if err != nil {
 		return err
 	}
 
-	clusterClient, err := clusterv1client.NewForConfig(a.config)
+	clusterClient, err := clusterv1client.NewForConfig(config)
 	if err != nil {
 		return err
 	}
@@ -181,56 +162,14 @@ func (a *cloudeventsAddonManager) Start(ctx context.Context) error {
 		return err
 	}
 
-	err = a.startWithInformers(ctx, workClient, workInformers, kubeInformers, addonInformers, clusterInformers, dynamicInformers)
-	if err != nil {
-		return err
-	}
-
-	kubeInformers.Start(ctx.Done())
-	go workInformers.Informer().Run(ctx.Done())
-	addonInformers.Start(ctx.Done())
-	clusterInformers.Start(ctx.Done())
-	dynamicInformers.Start(ctx.Done())
-	return nil
-}
-
-func (a *cloudeventsAddonManager) StartWithInformers(ctx context.Context,
-	kubeInformers kubeinformers.SharedInformerFactory,
-	workInformers workinformers.SharedInformerFactory,
-	addonInformers addoninformers.SharedInformerFactory,
-	clusterInformers clusterv1informers.SharedInformerFactory,
-	dynamicInformers dynamicinformer.DynamicSharedInformerFactory) error {
-	// TODO: Implement this method to support usage within the addon template.
-	// Requires updating sdk-go to enable clientHolder to manage the shared informer for ManifestWork.
-	return fmt.Errorf("method StartWithInformers is not implemented")
-}
-
-func (a *cloudeventsAddonManager) startWithInformers(ctx context.Context,
-	workClient workclientset.Interface,
-	workInformers workv1informers.ManifestWorkInformer,
-	kubeInformers kubeinformers.SharedInformerFactory,
-	addonInformers addoninformers.SharedInformerFactory,
-	clusterInformers clusterv1informers.SharedInformerFactory,
-	dynamicInformers dynamicinformer.DynamicSharedInformerFactory) error {
-
-	kubeClient, err := kubernetes.NewForConfig(a.config)
-	if err != nil {
-		return err
-	}
-
-	addonClient, err := addonv1alpha1client.NewForConfig(a.config)
-	if err != nil {
-		return err
-	}
-
 	v1CSRSupported, v1beta1Supported, err := utils.IsCSRSupported(kubeClient)
 	if err != nil {
 		return err
 	}
 
-	for _, agentImpl := range a.addonAgents {
+	for _, agentImpl := range addonAgents {
 		for _, configGVR := range agentImpl.GetAgentAddonOptions().SupportedConfigGVRs {
-			a.addonConfigs[configGVR] = true
+			addonConfigs[configGVR] = true
 		}
 	}
 
@@ -240,14 +179,14 @@ func (a *cloudeventsAddonManager) startWithInformers(ctx context.Context,
 		clusterInformers.Cluster().V1().ManagedClusters(),
 		addonInformers.Addon().V1alpha1().ManagedClusterAddOns(),
 		workInformers,
-		a.addonAgents,
+		addonAgents,
 	)
 
 	registrationController := registration.NewAddonRegistrationController(
 		addonClient,
 		clusterInformers.Cluster().V1().ManagedClusters(),
 		addonInformers.Addon().V1alpha1().ManagedClusterAddOns(),
-		a.addonAgents,
+		addonAgents,
 	)
 
 	// This controller is used during migrating addons to be managed by addon-manager.
@@ -256,26 +195,26 @@ func (a *cloudeventsAddonManager) startWithInformers(ctx context.Context,
 	managementAddonController := cmamanagedby.NewCMAManagedByController(
 		addonClient,
 		addonInformers.Addon().V1alpha1().ClusterManagementAddOns(),
-		a.addonAgents,
-		utils.FilterByAddonName(a.addonAgents),
+		addonAgents,
+		utils.FilterByAddonName(addonAgents),
 	)
 
 	var addonConfigController, managementAddonConfigController factory.Controller
-	if len(a.addonConfigs) != 0 {
+	if len(addonConfigs) != 0 {
 		addonConfigController = addonconfig.NewAddonConfigController(
 			addonClient,
 			addonInformers.Addon().V1alpha1().ManagedClusterAddOns(),
 			addonInformers.Addon().V1alpha1().ClusterManagementAddOns(),
 			dynamicInformers,
-			a.addonConfigs,
-			utils.FilterByAddonName(a.addonAgents),
+			addonConfigs,
+			utils.FilterByAddonName(addonAgents),
 		)
 		managementAddonConfigController = cmaconfig.NewCMAConfigController(
 			addonClient,
 			addonInformers.Addon().V1alpha1().ClusterManagementAddOns(),
 			dynamicInformers,
-			a.addonConfigs,
-			utils.FilterByAddonName(a.addonAgents),
+			addonConfigs,
+			utils.FilterByAddonName(addonAgents),
 		)
 	}
 
@@ -292,14 +231,14 @@ func (a *cloudeventsAddonManager) startWithInformers(ctx context.Context,
 			kubeInformers.Certificates().V1().CertificateSigningRequests(),
 			nil,
 			addonInformers.Addon().V1alpha1().ManagedClusterAddOns(),
-			a.addonAgents,
+			addonAgents,
 		)
 		csrSignController = certificate.NewCSRSignController(
 			kubeClient,
 			clusterInformers.Cluster().V1().ManagedClusters(),
 			kubeInformers.Certificates().V1().CertificateSigningRequests(),
 			addonInformers.Addon().V1alpha1().ManagedClusterAddOns(),
-			a.addonAgents,
+			addonAgents,
 		)
 	} else if v1beta1Supported {
 		csrApproveController = certificate.NewCSRApprovingController(
@@ -308,11 +247,11 @@ func (a *cloudeventsAddonManager) startWithInformers(ctx context.Context,
 			nil,
 			kubeInformers.Certificates().V1beta1().CertificateSigningRequests(),
 			addonInformers.Addon().V1alpha1().ManagedClusterAddOns(),
-			a.addonAgents,
+			addonAgents,
 		)
 	}
 
-	a.syncContexts = append(a.syncContexts, deployController.SyncContext())
+	a.SetSyncContexts(append(syncContexts, deployController.SyncContext()))
 
 	go deployController.Run(ctx, 1)
 	go registrationController.Run(ctx, 1)
@@ -330,19 +269,21 @@ func (a *cloudeventsAddonManager) startWithInformers(ctx context.Context,
 	if csrSignController != nil {
 		go csrSignController.Run(ctx, 1)
 	}
+
+	kubeInformers.Start(ctx.Done())
+	go workInformers.Informer().Run(ctx.Done())
+	addonInformers.Start(ctx.Done())
+	clusterInformers.Start(ctx.Done())
+	dynamicInformers.Start(ctx.Done())
 	return nil
 }
 
 // New returns a new addon manager with the given config and optional options
 func New(config *rest.Config, opts *CloudEventsOptions) (addonmanager.AddonManager, error) {
 	cloudeventsAddonManager := &cloudeventsAddonManager{
-		config:       config,
-		syncContexts: []factory.SyncContext{},
-		addonConfigs: map[schema.GroupVersionResource]bool{},
-		addonAgents:  map[string]agent.AgentAddon{},
+		BaseAddonManagerImpl: addonmanager.NewBaseAddonManagerImpl(config),
+		options:              opts,
 	}
-	// set options from the given options
-	cloudeventsAddonManager.options = opts
 
 	return cloudeventsAddonManager, nil
 }
